Homework-4-3-1: add check if event already started

+++ sports-betting-application-spring-app/src/main/java/com/epam/training/sportsbetting/service/impl/WagerServiceImpl.java
@@ -7,7 +7,9 @@ import org.springframework.stereotype.Service;
 
 import com.epam.training.sportsbetting.domain.Wager;
 import com.epam.training.sportsbetting.domain.user.User;
+import com.epam.training.sportsbetting.exception.EventNotStartedYetException;
 import com.epam.training.sportsbetting.repository.WagerRepository;
+import com.epam.training.sportsbetting.service.SportEventService;
 import com.epam.training.sportsbetting.service.WagerService;
 import com.google.common.collect.Lists;
 
@@ -17,6 +19,8 @@ public class WagerServiceImpl implements WagerService {
 
     @Autowired
     private WagerRepository wagerRepository;
+    @Autowired
+    private SportEventService sportEventService;
 
     @Override
     public List<Wager> findAll() {
@@ -39,8 +43,16 @@ public class WagerServiceImpl implements WagerService {
     }
 
     @Override
-    public void deleteById(Long idWager) {
-        wagerRepository.deleteById(idWager);
+    public void deleteById(Long idWager) throws EventNotStartedYetException {
+        if (eventNotStartedYet(idWager)) {
+            wagerRepository.deleteById(idWager);
+        } else {
+            throw new EventNotStartedYetException("Oops! U are late :) ");
+        }
+    }
+
+    private boolean eventNotStartedYet(Long idWager) {
+        return sportEventService.findByWager(findById(idWager)).isPresent();
     }
 
 }
